% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{cds_assign_star}
\alias{cds_assign_star}
\alias{cds_remove_star}
\title{Assign or remove stars to/from datasets}
\usage{
cds_assign_star(dataset, ..., token = cds_get_token())

cds_remove_star(dataset, ..., token = cds_get_token())
}
\arguments{
\item{dataset}{TODO}

\item{...}{Ignored}

\item{token}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
#TODO
}
