% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{cds_datasets}
\alias{cds_datasets}
\title{List or search Climate Data Service datasets}
\usage{
cds_datasets(search = NULL, page = 0, limit = 50, ...)
}
\arguments{
\item{search}{TODO}

\item{page}{TODO}

\item{limit}{TODO}

\item{...}{Ignored}
}
\value{
TODO
}
\description{
TODO
}
\examples{
if (interactive()) {
  #TODO
}
}
