% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{cds_get_account}
\alias{cds_get_account}
\title{Get account details}
\usage{
cds_get_account(token = cds_get_token(), ...)
}
\arguments{
\item{token}{An API key to be used for authentication. Will use
\code{cds_get_token()} by default.}

\item{...}{Ignored}
}
\value{
Returns a named \code{list} with account details.
}
\description{
Retrieve account details associated with the provided \code{token}
}
\examples{
if (interactive() && cds_token_works()) {
  cds_get_account()
}
}
\seealso{
Other authentication-functions:
\code{\link[=cds_account_metrics]{cds_account_metrics()}},
\code{\link[=cds_check_authentication]{cds_check_authentication()}},
\code{\link[=cds_get_token]{cds_get_token()}}
}
\concept{authentication-functions}
