% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{cds_list_collections}
\alias{cds_list_collections}
\title{List Climate Data Service collections}
\usage{
cds_list_collections(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
TODO
}
\description{
TODO
}
\examples{
if (interactive()) {
  cds_collections()
}
}
