% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{cds_open_api}
\alias{cds_open_api}
\title{List information about the API used for the catalogue}
\usage{
cds_open_api(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
TODO
}
\description{
TODO
}
\examples{
if (interactive()) {
  cds_open_api()
}
}
