% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{cds_retrieve}
\alias{cds_retrieve}
\title{TODO retrieve split this in submit and download functions}
\usage{
cds_retrieve(
  id,
  variable,
  product_type,
  year,
  month,
  day,
  pressure_level,
  format,
  ...,
  token = cds_get_token()
)
}
\arguments{
\item{id}{TODO}

\item{variable}{TODO}

\item{product_type}{TODO}

\item{year}{TODO}

\item{month}{TODO}

\item{day}{TODO}

\item{pressure_level}{TODO}

\item{format}{TODO}

\item{...}{Ignored}

\item{token}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
if (interactive() && cds_token_works()) {
  cds_retrieve(
    id             = "reanalysis-era5-pressure-levels",
    variable       = "geopotential",
    product_type   = "reanalysis",
    year           = "2024",
    month          = "03",
    day            = "01",
    pressure_level = "1000",
    format         = "netcdf"
  )
}
}
