% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{cds_set_token}
\alias{cds_set_token}
\title{Store a Climate Data Service API key for the current R session}
\usage{
cds_set_token(token, method = c("option", "sysenv"), ...)
}
\arguments{
\item{token}{TODO}

\item{method}{TODO}

\item{...}{Ignored}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO

}
