% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopernicusClimate-package.R
\docType{package}
\name{CopernicusClimate-package}
\alias{CopernicusClimate}
\alias{CopernicusClimate-package}
\title{CopernicusClimate: Search Download and Handle Data from Copernicus Climate Data Service}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Subset and download data from EU Copernicus Climate Data Service: \url{https://cds.climate.copernicus.eu/}. Import information about the Earth's past, present and future climate from Copernicus into R without the need of external software.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pepijn-devries.github.io/CopernicusClimate/}
  \item \url{https://github.com/pepijn-devries/CopernicusClimate/}
  \item Report bugs at \url{https://github.com/pepijn-devries/CopernicusClimate/issues}
}

}
\author{
\strong{Maintainer}: Pepijn de Vries \email{pepijn.devries@outlook.com} (\href{https://orcid.org/0000-0002-7961-6646}{ORCID})

Authors:
\itemize{
  \item Pepijn de Vries \email{pepijn.devries@outlook.com} (\href{https://orcid.org/0000-0002-7961-6646}{ORCID})
}

}
\keyword{internal}
