% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licenses.R
\name{cds_accepted_licences}
\alias{cds_accepted_licences}
\title{List accepted licences}
\usage{
cds_accepted_licences(
  scope = c("None", "all", "dataset", "portal"),
  ...,
  token = cds_get_token()
)
}
\arguments{
\item{scope}{TODO}

\item{...}{Ignored}

\item{token}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
if (interactive() && cds_token_works()) {
  cds_accepted_licences("portal")
}
}
