% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{cds_account_metrics}
\alias{cds_account_metrics}
\title{Get Prometheus metrics for account}
\usage{
cds_account_metrics(token = cds_get_token(), ...)
}
\arguments{
\item{token}{TODO}

\item{...}{Ignored}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO

}
