% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{cds_account_metrics}
\alias{cds_account_metrics}
\title{Get Prometheus metrics for account}
\usage{
cds_account_metrics(token = cds_get_token(), ...)
}
\arguments{
\item{token}{An API key to be used for authentication. Will use
\code{cds_get_token()} by default.}

\item{...}{Ignored}
}
\value{
Returns text that can be interpreted with \href{https://prometheus.io/}{Prometheus}
}
\description{
Obtain account metrics that can be interpreted with \href{https://prometheus.io/}{Prometheus}
}
\examples{
if (interactive() && cds_token_works()) {
  cds_account_metrics()
}
}
