% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{cds_build_request}
\alias{cds_build_request}
\title{Prepare a request for downloading a dataset}
\usage{
cds_build_request(dataset, ...)
}
\arguments{
\item{dataset}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
if (interactive()) {
  cds_build_request(
    dataset        = "reanalysis-era5-pressure-levels",
    variable       = "geopotential",
    product_type   = "reanalysis",
    year           = "2024",
    month          = "03",
    day            = "01",
    pressure_level = "1000",
    format         = "netcdf"
  )
}
}
