% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{cds_catalogue_vocabulary}
\alias{cds_catalogue_vocabulary}
\title{List catalogue vocabulary}
\usage{
cds_catalogue_vocabulary(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
Returns a \code{data.frame} of keyword vocabulary used by the catalogue.
}
\description{
The catalogue uses a specific vocabulary for keywords. This function
produces an overview.
}
\examples{
if (interactive()) {
  cds_catalogue_vocabulary()
}
}
