% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{cds_check_authentication}
\alias{cds_check_authentication}
\alias{cds_token_works}
\title{Check if authentication works with a specific token}
\usage{
cds_check_authentication(token = cds_get_token(), ...)

cds_token_works(token = cds_get_token(), ...)
}
\arguments{
\item{token}{TODO}

\item{...}{Ignored}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO

}
