% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.R
\name{cds_cite_dataset}
\alias{cds_cite_dataset}
\title{Cite a dataset}
\usage{
cds_cite_dataset(dataset, ...)
}
\arguments{
\item{dataset}{The name of a dataset to be cited.}

\item{...}{Ignored}
}
\value{
Returns a \code{BibEntry}-class object, with citation details for the requested
dataset
}
\description{
Use this function to obtain citation details for a specific dataset
}
\examples{
if (interactive()) {
  cds_cite_dataset("reanalysis-era5-pressure-levels")
}
}
