% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{cds_dataset_form}
\alias{cds_dataset_form}
\title{Obtain list of options to subset a dataset}
\usage{
cds_dataset_form(dataset, ...)
}
\arguments{
\item{dataset}{TODO}

\item{...}{Ignored}
}
\value{
A named list with aspects of the dataset that can be subsetted, when defining
a job \emph{TODO} ref
}
\description{
TODO
}
