% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{cds_dataset_form}
\alias{cds_dataset_form}
\title{Obtain an overview of options to subset a dataset}
\usage{
cds_dataset_form(dataset, ...)
}
\arguments{
\item{dataset}{A name of a dataset to explore}

\item{...}{Ignored}
}
\value{
A \code{data.frame} with aspects of the dataset that can be subsetted, when defining
a job. See also \code{cds_submit_job()} and \code{cds_build_request()}
}
\description{
This function provides an overview of parameters that can be used to subset a
dataset. It can help you set up a request with \code{cds_submit_job()} or \code{cds_build_request()}.
}
\examples{
if (interactive() && cds_token_works()) {
  cds_dataset_form("reanalysis-era5-pressure-levels")
}
}
