% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{cds_delete_job}
\alias{cds_delete_job}
\title{Delete/cancel jobs submitted to the Climate Data Service}
\usage{
cds_delete_job(job_id, ..., token = cds_get_token())
}
\arguments{
\item{job_id}{TODO}

\item{...}{Ignored}

\item{token}{An API key to be used for authentication. Will use
\code{cds_get_token()} by default.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
#TODO
}
