% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{cds_download_jobs}
\alias{cds_download_jobs}
\title{Download specific jobs}
\usage{
cds_download_jobs(job_id, destination, names, ..., token = cds_get_token())
}
\arguments{
\item{job_id}{TODO}

\item{destination}{Destination path to store downloaded files}

\item{names}{File names for the downloaded files. If missing, it is taken from the job.}

\item{...}{Ignored}

\item{token}{An API key to be used for authentication. Will use
\code{cds_get_token()} by default.}
}
\value{
A \code{data.frame} of all downloaded files. Contains a column \code{local} with the path
to the locally stored files.
}
\description{
TODO
}
\examples{
# TODO

}
