% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{cds_estimate_costs}
\alias{cds_estimate_costs}
\title{Check the cost of a request against your quotum}
\usage{
cds_estimate_costs(dataset, ..., token = cds_get_token())
}
\arguments{
\item{dataset}{TODO}

\item{...}{TODO}

\item{token}{An API key to be used for authentication. Will use
\code{cds_get_token()} by default.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
if (interactive() && cds_token_works()) {
  cds_estimate_costs(
    dataset        = "reanalysis-era5-pressure-levels",
    variable       = "geopotential",
    product_type   = "reanalysis",
    year           = "2024",
    month          = "03",
    day            = "01",
    pressure_level = "1000",
    format         = "netcdf"
  )
}
}
