% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{cds_get_account}
\alias{cds_get_account}
\title{Get account details}
\usage{
cds_get_account(token = cds_get_token(), ...)
}
\arguments{
\item{token}{TODO}

\item{...}{Ignored}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO

}
