% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{cds_list_jobs}
\alias{cds_list_jobs}
\title{List jobs submitted to the Climate Data Service}
\usage{
cds_list_jobs(job_id = NULL, ..., token = cds_get_token())
}
\arguments{
\item{job_id}{TODO}

\item{...}{Ignored}

\item{token}{An API key to be used for authentication. Will use
\code{cds_get_token()} by default.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
if (interactive() && cds_token_works()) {
  cds_list_jobs()
}
}
