% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licenses.R
\name{cds_list_licences}
\alias{cds_list_licences}
\title{Obtain a list of licences that can be accepted}
\usage{
cds_list_licences(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
Returns a \code{data.frame} of available licences that can be accepted.
}
\description{
Some datasets require you to accept specific licences. This function will provide
an overview of all licences associated with datasets and can be accepted.
}
\examples{
if (interactive()) {
  cds_list_licences()
}
}
\seealso{
Other licences: 
\code{\link{cds_accept_licence}()},
\code{\link{cds_accepted_licences}()}
}
\concept{licences}
