% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{cds_starred}
\alias{cds_starred}
\title{List starred datasets}
\usage{
cds_starred(..., token = cds_get_token())
}
\arguments{
\item{...}{Ignored}

\item{token}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
if (interactive() && cds_token_works()) {
  cds_starred()
}
}
