% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{cds_submit_job}
\alias{cds_submit_job}
\title{Submit a download job for a dataset}
\usage{
cds_submit_job(
  dataset,
  ...,
  wait = TRUE,
  check_quotum = TRUE,
  check_licence = TRUE,
  token = cds_get_token()
)
}
\arguments{
\item{dataset}{TODO}

\item{...}{TODO}

\item{wait}{TODO}

\item{check_quotum}{TODO}

\item{check_licence}{TODO}

\item{token}{An API key to be used for authentication. Will use
\code{cds_get_token()} by default.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
if (interactive() && cds_token_works()) {
  job <- cds_submit_job(
      dataset        = "reanalysis-era5-pressure-levels",
      variable       = "geopotential",
      product_type   = "reanalysis",
      year           = "2024",
      month          = "03",
      day            = "01",
      pressure_level = "1000",
      format         = "netcdf"
    )
}
}
